requirements.txt
streamlit
google-generativeai>=0.8.3
python-dotenv
scikit-learn
numpy
pypdf
beautifulsoup4
requests
youtube-transcript-api
faiss-cpu
langchain
langchain-community
langchain-google-genai
flask
gunicorn